+++ test/mocks/ERC721Mock.sol
@@ -9,29 +9,40 @@ import {ERC721} from '@openzeppelin/contracts/token/ERC721/ERC721.sol';
  * This mock just provides a public safeMint, mint, and burn functions for testing purposes
  */
 contract ERC721Mock is ERC721 {
-    constructor(string memory name, string memory symbol) ERC721(name, symbol) {}
+    constructor(
+        string memory name,
+        string memory symbol
+    ) ERC721(name, symbol) {}
 
     function baseURI() public view returns (string memory) {
         return _baseURI();
     }
 
-    function mint(address to, uint256 tokenId) public {
+    function mint(
+        address to,
+        uint tokenId
+    ) public {
         _mint(to, tokenId);
     }
 
-    function safeMint(address to, uint256 tokenId) public {
+    function safeMint(
+        address to,
+        uint tokenId
+    ) public {
         _safeMint(to, tokenId);
     }
 
     function safeMint(
         address to,
-        uint256 tokenId,
+        uint tokenId,
         bytes memory _data
     ) public {
         _safeMint(to, tokenId, _data);
     }
 
-    function burn(uint256 tokenId) public {
+    function burn(
+        uint tokenId
+    ) public {
         _burn(tokenId);
     }
 }
